created cypher module refactored decrypt and encrypt have 100% cov on simple cov all tests passing
