+++ test/test_helper.rb
@@ -1,4 +1,4 @@
-require 'minitest/autorun'
-require 'minitest/pride'
 require 'simplecov'
 SimpleCov.start
+require 'minitest/autorun'
+require 'minitest/pride'
